spack.new_installer: split/group package
